Refactor Uptime components and update import paths for improved structure
- Deleted several unused components related to monitor management, including MonitorResponseTimeChart, MonitorsTable, and various dialog components, streamlining the codebase.
- Updated import paths for components to reflect the new structure, ensuring consistency across the application.
- Adjusted the UptimeLayout to import Sidebar from the correct path, enhancing modularity.
- Improved overall organization of the Uptime components for better maintainability and clarity.
