+++ client/src/app/uptime/monitors/components/Sidebar.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { cn } from "../../../lib/utils";
+import { cn } from "../../../../lib/utils";
 import { Activity, AlertCircle, Plug2, Globe } from "lucide-react";
 
 export function Sidebar() {
